+++ src/components/Toast/src/ToastLayer.vue
@@ -8,7 +8,7 @@
 		]"
 		:before-enter-class="$s.fade_bottom"
 	>
-		<render-fn
+		<toast-render-fn
 			v-for="toast in toastApi.state.toasts"
 			:key="toast.key"
 			:render-fn="toast.renderFn"
@@ -19,7 +19,7 @@
 <script>
 import Vue from 'vue';
 import { MTransitionStack } from '@square/maker/utils/TransitionStack';
-import RenderFn from './RenderFn';
+import ToastRenderFn from './ToastRenderFn';
 import toastApi from './toast-api';
 
 function randomKey() {
@@ -82,7 +82,7 @@ const apiMixin = {
 				}
 			},
 
-			// closeSelf() is implemented in ./RenderFn.js
+			// closeSelf() is implemented in ./ToastRenderFn.js
 		};
 
 		if (!this.toastApi) {
@@ -98,7 +98,7 @@ const apiMixin = {
 export default {
 	components: {
 		MTransitionStack,
-		RenderFn,
+		ToastRenderFn,
 	},
 
 	inject: {
